created authentication endpoints and updated readme

+++ controllers/clothingitem.js
@@ -3,6 +3,7 @@ const {
   BAD_REQUEST_CODE,
   DEFAULT_ERROR_CODE,
   NOT_FOUND,
+  FORBIDDEN,
 } = require('../utils/error');
 
 // get all clothing items
@@ -51,10 +52,25 @@ const createItem = (req, res) => {
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
-  Items.findByIdAndDelete(itemId)
+  const userId = req.user._id;
+
+  // Find the item to check ownership
+  Items.findById(itemId)
     .orFail()
-    // Successful deletion returns the deleted document
-    .then((item) => res.status(200).send({ data: item }))
+    .then((item) => {
+      // Check if the requesting user is the owner
+      if (item.owner.toString() !== userId) {
+        return res
+          .status(FORBIDDEN)
+          .send({ message: 'You do not have permission to delete this item' });
+      }
+      // Proceed to delete the item
+      return Items.findByIdAndDelete(itemId)
+        .then((deletedItem) => res.status(200).send({ data: deletedItem }))
+        .catch(() => res
+          .status(DEFAULT_ERROR_CODE)
+          .send({ message: `${DEFAULT_ERROR_CODE}: Server Error` }));
+    })
     .catch((err) => {
       // Invalid ID format
       if (err.name === 'CastError') {
